Moves errors to a component
